requirements.txt
python-telegram-bot==20.0
SpeechRecognition==3.8.1
pydub==0.25.1
gTTS==2.2.3
ffmpeg-python==0.2.0